dom/base/test/meta_viewport/test_meta_viewport_maximum_scale_0.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>meta viewport test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=device-width, minimum-scale=1, maximum-scale=0">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>width=device-width, minimum-scale=1, maximum-scale=0</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function test1() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      is(info.maxZoom,     1,   "maximum_scale should be clamped");
      is(info.defaultZoom, 1,   "initial scale should be 1");
      is(info.minZoom,     1,   "minimum scale should be 1");
      is(info.width,       800, "width should be 800");
      is(info.height,      480, "height should be 480");
    });
  </script>
</body>
</html>
